refactor: 엑셀 추출 함수 분리
